pyproject.toml
[project]
name = "lerobot_policy_a2"
version = "0.1.0"
description = "A2 Policy for LeRobot - ViLGP3D grasp and place policy using CLIP-based action selection"
readme = "README.md"
requires-python = ">= 3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Denis Grachev"}
]
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.0.0",
    "open3d>=0.17.0",
    "Pillow>=9.0.0",
    "huggingface_hub>=0.20.0",
    "ftfy",
    "regex",
    "tqdm",
    "packaging",
    "torchvision>=0.15.0",
    "pybullet>=3.2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lerobot_policy_a2"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
